requirements.txt
simple-mqtt
python-slugify
